Fix SIGINT handling when running without a GUI
When the UI isn't running, the Qt application spends all its time running in its
event loop (implemented in C).  During that time, we can't process any (Unix)
signals in Python.  So, in order to handle signals, we have to occationally
execute some Python code.  So, we use a QTimer to occationally call into Python
and do no work.