Speed up startup time
We were calling parameter_changed on presets a bajillion times because that
method was called every time a parameter was added to a preset.  We now only
call it once, when it is added to the playlist.  This is a bit of a hack and
probably means that the initialization model needs to be redesigned.