Use time.time() instead of time.clock()
time.clock() is CPU-time (on Unix platforms), while time.time() is wall clock
time. Therefore, time.time() is more appropriate for calculating how long to
sleep to maintain a steady frame rate.

This commit might also fix a bug I've been observing on the lotus where Firemix
seems to hang for a few minutes after running for a while and then resume. This
might be explainable by time.clock() overflowing on the BeagleBone's 32
processor and causing a long sleep time (see the NOTES section of the clock(3)
man page).